+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.69.1
 
+- Node (GSE): Implement toToken() and encryptUserId() for CallLinkSecretParams